added move function to picarhelper and added check for base corners
